change config
